+++ app/_common/store/usePositionStore.ts
@@ -1,7 +1,8 @@
 import { persist } from "zustand/middleware";
 import { create } from "zustand";
 
-interface PositionState {
+import getGeolocation from "@/app/_common/utils/getGeolocation";
+
 interface Coordinate {
   longitude: number;
   latitude: number;
@@ -23,22 +24,7 @@ const usePositionStore = create(
       longitude: 0,
       latitude: 0,
       setPosition: () => {
-        navigator.geolocation.getCurrentPosition((position) => {
-          const { longitude, latitude } = position.coords;
-          if (!longitude || !latitude) {
-            set({ longitude: 0, latitude: 0, isGPSOn: false });
-            errorCallback({
-              code: 2,
-              message: "",
-              PERMISSION_DENIED: 1,
-              POSITION_UNAVAILABLE: 2,
-              TIMEOUT: 3,
-            });
-            return;
-          }
-
-          set({ longitude, latitude, isGPSOn: true });
-        }, errorCallback, options);
+        getGeolocation(set);
       },
       getPosition: () => {
         const { longitude, latitude } = get();
@@ -50,27 +36,5 @@ const usePositionStore = create(
     },
   ),
 );
-
-const errorCallback = (error: GeolocationPositionError) => {
-  switch (error.code) {
-    case error.PERMISSION_DENIED:
-      console.error("사용자가 Geolocation API의 사용 요청을 거부했습니다");
-      break;
-    case error.POSITION_UNAVAILABLE:
-      console.error("가져온 위치 정보를 사용할 수 없습니다.");
-      break;
-    case error.TIMEOUT:
-      console.error("위치 정보를 가져오기 위한 요청이 허용 시간을 초과했습니다");
-      break;
-    default:
-      console.error("알 수 없는 오류가 발생했습니다.");
-      break;
-  }
-};
-
-const options = {
-  maximumAge: 300000,
-  timeout: 15000
-};
 
 export default usePositionStore;
